InvGamma {MCMCpack}	R Documentation

The Inverse Gamma Distribution

Description

Density function and random generation from the inverse gamma distribution.

Usage

dinvgamma(x, shape, scale = 1)

rinvgamma(n, shape, scale = 1)

Arguments

x
Scalar location to evaluate density.

shape
Scalar shape parameter.

scale
Scalar scale parameter (default value one).

n
Number of draws from the distribution.

Details

An inverse gamma random variable with shape a and scale b has mean \frac{b}{a-1} (assuming a>1) and variance \frac{b^2}{(a-1)^2(a-2)} (assuming a>2).

Value

dinvgamma evaluates the density at x.

rinvgamma takes n draws from the inverse Gamma distribution. The parameterization is consistent with the Gamma Distribution in the stats package.

References

Andrew Gelman, John B. Carlin, Hal S. Stern, and Donald B. Rubin. 2004. Bayesian Data Analysis. 2nd Edition. Boca Raton: Chapman & Hall.

See Also

GammaDist

Examples

density <- dinvgamma(4.2, 1.1)
draws <- rinvgamma(10, 3.2)

[Package MCMCpack version 1.4-6 Index]