InvWishart {MCMCpack}	R Documentation

The Inverse Wishart Distribution

Description

Density function and random generation from the Inverse Wishart distribution.

Usage

riwish(v, S)

diwish(W, v, S)

Arguments

v
Degrees of freedom (scalar).

S
Scale matrix (p \times p).

W
Positive definite matrix W (p \times p).

Details

The mean of an inverse Wishart random variable with v degrees of freedom and scale matrix S is (v-p-1)^{-1}S.

Value

diwish evaluates the density at positive definite matrix W. riwish generates one random draw from the distribution.

Examples

density <- diwish(matrix(c(2,-.3,-.3,4),2,2), 3, matrix(c(1,.3,.3,1),2,2))
draw <- riwish(3, matrix(c(1,.3,.3,1),2,2))

[Package MCMCpack version 1.4-6 Index]