Political Economic Risk Data from 62 Countries in 1987

Political Economic Risk Data from 62 Countries in 1987.

Format

A data frame with 62 observations on the following 9 variables. All data points are from 1987. See Quinn (2004) for more details.

country

a factor with levels Argentina through Zimbabwe

courts

an ordered factor with levels 0 < 1.courts is an indicator of whether the country in question is judged to have an independent judiciary. From Henisz (2002).

barb2

a numeric vector giving the natural log of the black market premium in each country. The black market premium is coded as the black market exchange rate (local currency per dollar) divided by the official exchange rate minus 1. From Marshall, Gurr, and Harff (2002).

prsexp2

an ordered factor with levels 0 < 1 < 2 < 3 < 4 < 5, giving the lack of expropriation risk. From Marshall, Gurr, and Harff (2002).

prscorr2

an ordered factor with levels 0 < 1 < 2 < 3 < 4 < 5, measuring the lack of corruption. From Marshall, Gurr, and Harff (2002).

gdpw2

a numeric vector giving the natural log of real GDP per worker in 1985 international prices. From Alvarez et al. (1999).

Source

Mike Alvarez, Jose Antonio Cheibub, Fernando Limongi, and Adam Przeworski. 1999. ``ACLP Political and Economic Database.''

Witold J. Henisz. 2002. ``The Political Constraint Index (POLCON) Dataset.''

Monty G. Marshall, Ted Robert Gurr, and Barbara Harff. 2002. ``State Failure Task Force Problem Set.''

References

Kevin M. Quinn. 2004. ``Bayesian Factor Analysis for Mixed Ordinal and Continuous Response.'' Political Analyis. 12: 338-353.