choicevar {MCMCpack}	R Documentation

Handle Choice-Specific Covariates in Multinomial Choice Models

Description

This function handles choice-specific covariates in multinomial choice models. See the example for an example of useage.

Usage

choicevar(var, varname, choicelevel)

Arguments

var
The is the name of the variable in the dataframe.

varname
The name of the new variable to be created.

choicelevel
The level of y that the variable corresponds to.

Value

The new variable used by the MCMCmnl() function.

See Also

MCMCmnl

[Package MCMCpack version 1.4-6 Index]