make.breaklist {MCMCpack}	R Documentation

Vector of break numbers

Description

This function generates a vector of break numbers using the output of testpanelSubjectBreak. The function performs a pairwise comparison of models using Bayes Factors.

Usage

make.breaklist(BF, threshold = 3)

Arguments

BF
output of testpanelSubjectBreak.

threshold
The Bayes Factor threshold to pick the best model. If a Bayes factor of two models is smaller than threshold, the model with a smaller number of break is chosen to avoid the over-identification problem. Users can change threshold into any positive number. The default value of 3 is chosen as it indicates the existence of "substantial evidence" in favor of the model in the numerator according to Jeffreys' scale.

Value

Vector fo break numbers.

References

Jong Hee Park, 2011. “A Unified Method for Dynamic and Cross-Sectional Heterogeneity: Introducing Hidden Markov Panel Models." Working Paper.

Harold Jeffreys, 1961. The Theory of Probability. Oxford University Press.

See Also

testpanelSubjectBreak

[Package MCMCpack version 1.4-6 Index]