mptable {MCMCpack}	R Documentation

Calculate the marginal posterior probabilities of predictors being included in a quantile regression model.

Description

This function extracts the marginal probability table produced by summary.qrssvs.

Usage

mptable(qrssvs)

Arguments

qrssvs
An object of class qrssvs. Typically this will be the gamma component of the list returned by SSVSquantreg.

Value

A table with the predictors listed together with their posterior marginal posterior probability of inclusion.

Author(s)

Craig Reed

See Also

SSVSquantreg

Examples

## Not run: 
set.seed(1)
epsilon<-rnorm(100)
set.seed(2)
x<-matrix(rnorm(1000),100,10)
y<-x[,1]+x[,10]+epsilon
qrssvs<-SSVSquantreg(y~x)
mptable(qrssvs$gamma)

## End(Not run)

[Package MCMCpack version 1.4-6 Index]