plot.qrssvs {MCMCpack}	R Documentation

Plot output from quantile regression stochastic search variable selection (QR-SSVS).

Description

This function produces a Trellis plot of the predictors on the y-axis versus the marginal posterior probability of inclusion on the x-axis.

Usage

## S3 method for class 'qrssvs'
plot(x, ...)

Arguments

x
An object of class qrssvs. Typically this will be the gamma component of the list returned by SSVSquantreg.

...
Further arguments

Value

An object with class "trellis". The associated update and print methods are documented in the "Lattice" package.

Author(s)

Craig Reed

References

Deepayan Sarkar. 2008. lattice: Lattice Graphics. R package version 0.17-17

See Also

SSVSquantreg, mptable, Lattice for a brief introduction to lattice displays and links to further documentation.

Examples

## Not run: 
set.seed(1)
epsilon<-rnorm(100)
set.seed(2)
x<-matrix(rnorm(1000),100,10)
y<-x[,1]+x[,10]+epsilon
qrssvs<-SSVSquantreg(y~x)
plot(qrssvs$gamma)
## Modify the graph by increasing the fontsize on the axes
qrssvsplot<-plot(qrssvs$gamma)
update(qrssvsplot, scales=list(cex=3))

## End(Not run)

[Package MCMCpack version 1.4-6 Index]