plotChangepoint {MCMCpack}	R Documentation

Posterior Density of Regime Change Plot

Description

Plot the posterior density of regime change.

Usage

plotChangepoint(mcmcout,
  main = "Posterior Density of Regime Change Probabilities",
  xlab = "Time", ylab = "", verbose = FALSE, start = 1,
  overlay = FALSE)

Arguments

mcmcout
The mcmc object containing the posterior density sample from a changepoint model. Note that this must have a

prob.state attribute.

main
Title of the plot

xlab
Label for the x-axis.

ylab
Label for the y-axis.

verbose
If verbose=TRUE, expected changepoints are printed.

start
The time of the first observation to be shown in the time series plot.

overlay
If overlay=TRUE, the probability of each regime change is drawn separately, which will be useful to draw multiple plots in one screen. See the example in MCMCpoissonChange. Otherwise, multiple plots of regime change probabilities will be drawn.

See Also

MCMCpoissonChange, MCMCbinaryChange

[Package MCMCpack version 1.4-6 Index]