plotHDPChangepoint {MCMCpack}	R Documentation

Posterior Changepoint Probabilities from HDP-HMM

Description

Plot the posterior density of regime change.

Usage

plotHDPChangepoint(mcmcout, main = "Posterior Changepoint Probabilities",
  xlab = "Time", ylab = "", start = 1)

Arguments

mcmcout
The mcmc object containing the posterior density sample from a changepoint model. Note that this must be from a HDP-HMM sampler.

main
Title of the plot

xlab
Label for the x-axis.

ylab
Label for the y-axis.

start
The time of the first observation to be shown in the time series plot.

See Also

HDPHMMpoisson, HDPHMMnegbin, HDPHSMMnegbin

[Package MCMCpack version 1.4-6 Index]