plotState {MCMCpack}	R Documentation

Changepoint State Plot

Description

Plot the posterior probability that each time point is in each state.

Usage

plotState(mcmcout, main = "Posterior Regime Probability",
  ylab = expression(paste("Pr(", S[t], "= k |", Y[t], ")")),
  legend.control = NULL, cex = 0.8, lwd = 1.2, start = 1)

Arguments

mcmcout
The mcmc object containing the posterior density sample from a changepoint model. Note that this must have a prob.state attribute.

main
Title of the plot.

ylab
Label for the y-axis.

legend.control
Control the location of the legend. It is necessary to pass both the x and y locations; i.e., c(x,y).

cex
Control point size.

lwd
Line width parameter.

start
The time of the first observation to be shown in the time series plot.

See Also

MCMCpoissonChange, MCMCbinaryChange

[Package MCMCpack version 1.4-6 Index]