procrustes {MCMCpack}	R Documentation

Procrustes Transformation

Description

This function performs a Procrustes transformation on a matrix X to minimize the squared distance between X and another matrix Xstar.

Usage

procrustes(X, Xstar, translation = FALSE, dilation = FALSE)

Arguments

X
The matrix to be transformed.

Xstar
The target matrix.

translation
logical value indicating whether X should be translated.

dilation
logical value indicating whether X should be dilated.

Details

R, tt, and s are chosen so that:

s X R + 1 tt' \approx X^*

X.new is given by:

X_{new} = s X R + 1 tt'

Value

A list containing: X.new the matrix that is the Procrustes transformed version of X, R the rotation matrix, tt the translation vector, and s the scale factor.

References

Borg and Groenen. 1997. Modern Multidimensional Scaling. New York: Springer. pp. 340-342.

See Also

MCMCirtKd

[Package MCMCpack version 1.4-6 Index]