read.Scythe {MCMCpack}	R Documentation

Read a Matrix from a File written by Scythe

Description

This function reads a matrix from an ASCII file in the form produced by the Scythe Statistical Library. Scythe output files contain the number of rows and columns in the first row, followed by the data.

Usage

read.Scythe(infile = NA)

Arguments

infile
The file to be read. This can include path information.

Value

A matrix containing the data stored in the read file.

References

Daniel Pemstein, Kevin M. Quinn, and Andrew D. Martin. 2007. Scythe Statistical Library 1.0. http://scythe.lsa.umich.edu.

See Also

write.Scythe

Examples

  ## Not run: 
  mymatrix <- read.Scythe("myfile.txt")
  
## End(Not run)

[Package MCMCpack version 1.4-6 Index]