summaryqrssvs {MCMCpack}	R Documentation

Summarising the results of quantile regression stochastic search variable selection (QR-SSVS).

Description

This function produces a table of predictors and their associated marginal posterior probability of inclusion. It also returns the median probability model (see the details section).

Usage

## S3 method for class 'qrssvs'
summary(object, ...)

Arguments

object
An object of class qrssvs. Typically this will be the gamma component of the list returned by SSVSquantreg.

...
Further arguments.

Details

The median probability model is defined to be the model that contains any predictor with marginal posterior probability greater than or equal to 0.5. If the goal is to select a single model e.g. for prediction, Barbieri and Berger (2004) recommend the median probability model. In some cases, this will coincide with the maximum probability model.

Author(s)

Craig Reed

References

Maria M. Barbieri, and James O. Berger (2004). "Optimal predictive model selection". Annals of Statistics, 32, 870-897.

See Also

SSVSquantreg, mptable, topmodels

Examples

## Not run: 
set.seed(1)
epsilon<-rnorm(100)
set.seed(2)
x<-matrix(rnorm(1000),100,10)
y<-x[,1]+x[,10]+epsilon
qrssvs<-SSVSquantreg(y~x)
summary(qrssvs$gamma)

## End(Not run)

[Package MCMCpack version 1.4-6 Index]