write.Scythe {MCMCpack}	R Documentation

Write a Matrix to a File to be Read by Scythe

Description

This function writes a matrix to an ASCII file that can be read by the Sycthe Statistical Library. Scythe requires that input files contain the number of rows and columns in the first row, followed by the data.

Usage

write.Scythe(outmatrix, outfile = NA, overwrite = FALSE)

Arguments

outmatrix
The matrix to be written to a file.

outfile
The file to be written. This can include path information.

overwrite
A logical that determines whether an existing file should be over-written. By default, it protects the user from over-writing existing files.

Value

A zero if the file is properly written.

References

Daniel Pemstein, Kevin M. Quinn, and Andrew D. Martin. 2007. Scythe Statistical Library 1.0. http://scythe.lsa.umich.edu.

See Also

write.Scythe

Examples

  ## Not run: 
  write.Scythe(mymatrix, file.path(tempdir(), "myfile.txt"))
  
## End(Not run)

[Package MCMCpack version 1.4-6 Index]