xpnd {MCMCpack}	R Documentation

Expand a Vector into a Symmetric Matrix

Description

This function takes a vector of appropriate length (typically created using vech) and creates a symmetric matrix.

Usage

xpnd(x, nrow = NULL)

Arguments

x
A list of elements to expand into symmetric matrix.

nrow
The number of rows (and columns) in the returned matrix. Look into the details.

Details

This function is particularly useful when dealing with variance covariance matrices. Note that R stores matrices in column major order, and that the items in x will be recycled to fill the matrix if need be.

The number of rows can be specified or automatically computed from the number of elements in a given object via (-1 + √{(1 + 8 * length(x))}) / 2.

Value

An (nrows \times nrows) symmetric matrix.

See Also

vech

Examples

  xpnd(c(1,2,3,4,4,5,6,7,8,9),4)
  xpnd(c(1,2,3,4,4,5,6,7,8,9))

[Package MCMCpack version 1.4-6 Index]